Adds user onboarding call on initial root page
